+++ reservation-service/src/main/webapp/WEB-INF/views/index.jsp
@@ -6,7 +6,6 @@
 <!DOCTYPE html>
 <html>
 <head>
-
 <meta charset="utf-8">
 <title>Category</title>
 </head>
@@ -17,8 +16,13 @@
 
 	<div id = 'addReply'>
 		<form method="post" action="/category/create">
-			name : <input type="text" name="name"><br>
-			<input type="submit" value="확인">
+			<div>
+				<label for="name">NAME : </label>
+				<input type="text" name="name"><br>
+			</div>
+			<div>
+				<input type="submit" value="확인">
+			</div>
 		</form>
 	</div>
 
@@ -36,55 +40,38 @@
 					<td>${categoryVO.id}</td>
 					<td>${categoryVO.name}</td>
 					<td>
-						<form method="post" action="/category/modify">
-							 <input type="hidden" name="id" value="${categoryVO.id}" >
-							 수정 : <input type="text" name="name"><br>
-							<button value="수정" class="modify"> </button>
+						<form action="/category/modify" method="post" >
+							<div>
+								<input type="hidden" name="id" value="${categoryVO.id}">
+							</div>
+							<div>
+								<input type="text" name="name" placeholder="NEW NAME"><br>
+							</div>
+							<div class="button">
+								<button value="수정" name="newName"class="modify">MODIFY</button>
+							</div>
 						</form>
 					</td>
 					<td>
-						<button value="삭제" class="destory"> </button>
+						<div class="button">
+							<button value="삭제" class="destory"> REMOVE </button>
+						</div>
 					</td>
 				</tr>
 			</c:forEach>
 		</table>
 	</div>
-	
-<!-- 	<div id = 'modifyReply'>
-		<div class = 'modify-title'>
-			<div>
-				<input type='text' id='replytext'>
-			</div>				
-			<div>
-				<button type="button" id="modifyBtn"> MODIFY </button>
-				<button type="button" id="cancleBtn"> CANCLE </button>
-			</div>
-		</div>
-	</div> -->
-<script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.1.1.min.js"> </script>
+	<script src="resources/js/category.js"></script>
+	<script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.1.1.min.js"> </script>
 </body>
 </html>
 
 <script>
-
-function remove_category_view(view, id) {
-	view.remove();
-}
-
-function remove_category(view, id, callback) { 
-	$.ajax({
-		type:'delete',
-		url:'/categories/' + id,
-		dataType:'text', 
-		success:function() {
-			callback(view, id);
-		}
+$(function(){
+	$(document).on("click", ".destory", function(event){
+		var view = $(event.target).closest('.category');
+		var id = view.attr('id');
+		remove_category(view, id, remove_category_view);
 	});
-}
-
- $(document).on("click", ".destory", function(event){
-	var view = $(event.target).closest('.category');
-	var id = view.attr('id');
-	remove_category(view, id, remove_category_view);
 });
  </script>
