+++ reservation-service/src/main/webapp/resources/js/category.js
@@ -1,0 +1,15 @@
+function remove_category_view(view, id) {
+	view.remove();
+}
+
+
+function remove_category(view, id, callback) { 
+	$.ajax({
+		type:'post',
+		url:'/category/remove',
+		data:"id="+id,
+		success:function() {
+			callback(view, id);
+		}
+	});
+}
